+++ src/django_typer/management/commands/shellcompletion.py
@@ -37,11 +37,16 @@ from typer import Argument, Option
 from typer._click.core import ParameterSource
 from typer._click.globals import get_current_context
 from typer._click.shell_completion import CompletionItem, split_arg_string
-from typer.main import get_command as get_typer_command
 
 from django_typer.completers import these_strings
 from django_typer.completers.path import import_paths
-from django_typer.management import TyperCommand, command, get_command, initialize
+from django_typer.management import (
+    TyperCommand,
+    command,
+    get_command,
+    get_typer_command,
+    initialize,
+)
 from django_typer.shells import _completers
 from django_typer.types import COMMON_PANEL
 from django_typer.utils import detect_shell, get_usage_script, get_win_shell